feat: Add GitHub token setting for auto-updates
Introduces a new setting to store a GitHub personal access token. This token will be used by the auto-updater to securely fetch release information from GitHub, especially for private repositories or when rate limits are a concern.

The setting is added to the application's settings management, persisted, and loaded. It's also made available for configuring the auto-updater in Electron.